Impl Debug for TileGrids and make ::len() return usize

+++ src/graphics/geometry/hex.rs
@@ -39,7 +39,7 @@ fn tile_hit_test(p: impl Into<Point>) -> TileHit {
     TileHit::Inside
 }
 
-#[derive(Debug)]
+#[derive(Clone, Debug)]
 pub struct TileGrid {
     // Position in screen coordinates.
     // Tile at `pos` will be mapped to screen at screen_pos.
@@ -58,8 +58,8 @@ pub struct TileGrid {
 }
 
 impl TileGrid {
-    pub fn len(&self) -> i32 {
-        self.width * self.height
+    pub fn len(&self) -> usize {
+        (self.width * self.height) as usize
     }
 
     pub fn screen_pos(&self) -> Point {
